[worker] refactored test_enclave, now have a module with separate test files for ecall tests and integration tests
