+++ worker/src/tests/commons.rs
@@ -1,0 +1,15 @@
+use sgx_types::sgx_sha256_hash_t;
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Message {
+	pub account: String,
+	pub amount: u32,
+	pub sha256: sgx_sha256_hash_t
+}
+
+pub fn from_slice(bytes: &[u8]) -> [u8; 32] {
+	let mut array = [0; 32];
+	let bytes = &bytes[..array.len()]; // panics if not enough data
+	array.copy_from_slice(bytes);
+	array
+}
